HTTPレスポンス構造体を生成する処理を、Handler層に移行する
