+++ handler/book.go
@@ -6,6 +6,7 @@ import (
 	"github.com/ryota1116/stacked_books/domain/model/googleBooksApi"
 	"github.com/ryota1116/stacked_books/handler/http/request"
 	httpResponse "github.com/ryota1116/stacked_books/handler/http/response"
+	"github.com/ryota1116/stacked_books/handler/http/response/book"
 	"github.com/ryota1116/stacked_books/usecase"
 	"io/ioutil"
 	"net/http"
@@ -37,7 +38,8 @@ func (bh bookHandler) SearchBooks(w http.ResponseWriter, r *http.Request)  {
 	}
 
 	// リクエストボディのバリデーション
-	isValid, errMsg := request.BookHandlerFormValidator{GoogleBooksApiRequestBody: requestParameter}.Validate()
+	isValid, errMsg := request.BookHandlerFormValidator{
+		GoogleBooksApiRequestBody: requestParameter}.Validate()
 	if !isValid {
 		// クライアントにHTTPレスポンスを返す
 		response := httpResponse.Response{
@@ -49,17 +51,22 @@ func (bh bookHandler) SearchBooks(w http.ResponseWriter, r *http.Request)  {
 	}
 
 	// 外部APIで書籍を検索
-	searchBooksResult, err := bh.bookUseCase.SearchBooks(requestParameter)
+	responseFromGoogleBooksAPI, err := bh.bookUseCase.SearchBooks(requestParameter)
 	// 外部APIリクエストでエラーが発生した場合
 	if err != nil {
 		httpResponse.Return500Response(w, errors.New("検索に失敗しました"))
 		return
 	}
 
+	// GoogleBooksAPIのJSONレスポンスの構造体から、 書籍検索用のHTTPレスポンスボディ構造体を生成する
+	searchBooksResponse := book.SearchBooksResponseGenerator{
+		ResponseBodyFromGoogleBooksAPI: responseFromGoogleBooksAPI,
+	}.Execute()
+
 	// 正常なレスポンス
 	response := httpResponse.Response{
 		StatusCode:   http.StatusOK,
-		ResponseBody: searchBooksResult,
+		ResponseBody: searchBooksResponse,
 	}
 	response.ReturnResponse(w)
 }
