+++ handler/http/response/book/search_books_response_generator.go
@@ -1,9 +1,28 @@
-package googleBooksApi
+package book
+
+import "github.com/ryota1116/stacked_books/domain/model/googleBooksApi"
 
 // SearchBooksResponseGenerator : 書籍検索用レスポンスボディのジェネレーター
 type SearchBooksResponseGenerator struct {
-	ResponseBodyFromGoogleBooksAPI ResponseBodyFromGoogleBooksAPI `json:"response_body_from_google_books_api"`
+	ResponseBodyFromGoogleBooksAPI googleBooksApi.ResponseBodyFromGoogleBooksAPI `json:"response_body_from_google_books_api"`
 }
+
+type SearchBooksResponses []SearchBooksResponse
+
+// SearchBooksResponse : 書籍検索用のレスポンスボディ構造体。
+// GoogleBooksAPIを叩いた時に取得したJSONレスポンスのうち、
+// 必要なフィールドだけをセットしたレスポンスボディの構造体。
+type SearchBooksResponse struct {
+	GoogleBooksId string    `json:"google_books_id"`
+	Title	string			`json:"title"`
+	Authors	[]string		`json:"authors"`
+	Description	string		`json:"description"`
+	Isbn10 string			`json:"isbn_10"`
+	Isbn13 string			`json:"isbn_13"`
+	PageCount int 			`json:"page_count"`
+	RegisteredAt string	    `json:"created_at"`
+}
+
 
 // Execute : GoogleBooksAPIのJSONレスポンスの構造体から、
 // 必要なフィールドだけをセットした書籍検索用のレスポンスボディ構造体を生成する
