+++ usecase/book.go
@@ -5,7 +5,7 @@ import (
 )
 
 type BookUseCaseInterface interface {
-	SearchBooks(requestParameter googleBooksApi.RequestParameter) (googleBooksApi.SearchBooksResponses, error)
+	SearchBooks(requestParameter googleBooksApi.RequestParameter) (googleBooksApi.ResponseBodyFromGoogleBooksAPI, error)
 }
 
 type bookUseCase struct {
@@ -18,18 +18,14 @@ func NewBookUseCase() BookUseCaseInterface {
 }
 
 // SearchBooks : 外部APIを用いて書籍検索を行う
-func (bu bookUseCase) SearchBooks(requestParameter googleBooksApi.RequestParameter) (googleBooksApi.SearchBooksResponses, error) {
+func (bu bookUseCase) SearchBooks(requestParameter googleBooksApi.RequestParameter) (googleBooksApi.ResponseBodyFromGoogleBooksAPI, error) {
 	// 外部APIで書籍を検索
 	// 書籍検索用のレスポンスボディ構造体のスライス型
 	responseFromGoogleBooksAPI, err := googleBooksApi.Client{}.SendRequest(requestParameter.Title)
 	if err != nil {
-		return googleBooksApi.SearchBooksResponses{}, err
+		return responseFromGoogleBooksAPI, err
 	}
 
-	// GoogleBooksAPIのJSONレスポンスの構造体から、 書籍検索用のレスポンスボディ構造体を生成する
-	searchBooksResponse := googleBooksApi.SearchBooksResponseGenerator{
-		ResponseBodyFromGoogleBooksAPI: responseFromGoogleBooksAPI,
-	}.Execute()
-
-	return searchBooksResponse, nil
+	// GoogleBooksAPIのJSONレスポンスの構造体を返す
+	return responseFromGoogleBooksAPI, nil
 }
